+++ src/rules.jl
@@ -48,16 +48,38 @@ struct FuzzyOr{T <: AbstractFuzzyProposition, S <: AbstractFuzzyProposition} <:
 end
 Base.show(io::IO, fo::FuzzyOr) = print(io, '(', fo.left, " ∨ ", fo.right, ')')
 
+abstract type AbstractRule end
+
 """
 Describes a fuzzy implication rule IF antecedent THEN consequent.
 """
-struct FuzzyRule{T <: AbstractFuzzyProposition}
+struct FuzzyRule{T <: AbstractFuzzyProposition} <: AbstractRule
     "premise of the inference rule."
     antecedent::T
-    "consequences of the premise rule."
+    "consequences of the inference rule."
     consequent::Vector{FuzzyRelation}
 end
 Base.show(io::IO, r::FuzzyRule) = print(io, r.antecedent, " --> ", r.consequent...)
+
+"""
+Weighted fuzzy rule. In Mamdani systems, the result of implication is scaled by the weight.
+In Sugeno systems, the result of the antecedent is scaled by the weight.
+"""
+struct WeightedFuzzyRule{T <: AbstractFuzzyProposition, S <: Real} <: AbstractRule
+    "premise of the inference rule."
+    antecedent::T
+    "consequences of the inference rule."
+    consequent::Vector{FuzzyRelation}
+    "weight of the rule."
+    weight::S
+end
+
+function Base.show(io::IO, r::WeightedFuzzyRule)
+    print(io, r.antecedent, " --> ", r.consequent..., " (", r.weight, ")")
+end
+
+@inline scale(w, ::FuzzyRule) = w
+@inline scale(w, r::WeightedFuzzyRule) = w * r.weight
 
 # comparisons (for testing)
 
@@ -71,6 +93,11 @@ Base.:(==)(p1::AbstractFuzzyProposition, p2::AbstractFuzzyProposition) = false
 
 function Base.:(==)(r1::FuzzyRule, r2::FuzzyRule)
     r1.antecedent == r2.antecedent && r1.consequent == r1.consequent
+end
+
+function Base.:(==)(r1::WeightedFuzzyRule, r2::WeightedFuzzyRule)
+    r1.antecedent == r2.antecedent && r1.consequent == r1.consequent &&
+        r1.weight == r2.weight
 end
 
 # utilities